Extract express-graphql extension
