+++ index.js
@@ -4,5 +4,6 @@ module.exports = {
   createClient: require('./src/client.js'),
   prepareSpaceGraph: require('./src/prepare-space-graph.js'),
   createSchema: require('./src/schema.js'),
-  createUI: require('./src/ui.js')
+  createUI: require('./src/ui.js'),
+  helpers: require('./src/helpers')
 };
