+++ package.json
@@ -29,7 +29,7 @@
     "dump": "mkdirp dump && npm run dump-graph && npm run dump-schema",
     "dump-graph": "graphqlviz http://localhost:4000/graphql | dot -Tpng -o dump/graph.png",
     "dump-schema": "fetch-graphql-schema http://localhost:4000/graphql -r -o dump/schema.graphql",
-    "lint": "eslint 'index.js' 'src/*.js' 'test/*.js' 'demo/*.js' 'dev/*.js'",
+    "lint": "eslint 'index.js' 'src/**/*.js' 'test/*.js' 'demo/*.js' 'dev/*.js'",
     "test": "tape 'test/*.js'",
     "coverage": "istanbul cover tape 'test/*.js'",
     "codecov": "cat coverage/coverage.json | codecov"
