+++ src/helpers/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  expressGraphqlExtension: require('./express-graphql-extension.js')
+};
